Add test for a valid SystemJS template.
- breaks out golden file comparison into separate lib
- renames additional entry point to cover bug fixed in
  ab283fafa1a498c09efdfb96df7d10f69a94dbe2
