+++ internal/e2e/BUILD.bazel
@@ -1,0 +1,9 @@
+package(default_visibility = ["//internal:__subpackages__"])
+
+load("//internal/js_library:js_library.bzl", "js_library")
+
+js_library(
+    name = "check_lib",
+    srcs = ["check.js"],
+    module_name = "check",
+)
