+++ internal/e2e/rollup/BUILD.bazel
@@ -37,4 +37,7 @@ jasmine_node_test(
         ":bundle.min_debug.js",
     ],
     node_modules = "//internal/test:node_modules",
+    deps = [
+        "//internal/e2e:check_lib",
+    ],
 )
