+++ internal/e2e/rollup_code_splitting/BUILD.bazel
@@ -6,7 +6,7 @@ rollup_bundle(
         ["*.js"],
         exclude = ["*.spec.js"],
     ),
-    additional_entry_points = ["internal/e2e/rollup_code_splitting/main2.js"],
+    additional_entry_points = ["internal/e2e/rollup_code_splitting/additional_entry.js"],
     entry_point = "internal/e2e/rollup_code_splitting/main1.js",
     node_modules = "//internal/test:node_modules",
 )
@@ -14,6 +14,14 @@ rollup_bundle(
 jasmine_node_test(
     name = "test",
     srcs = glob(["*.spec.js"]),
-    data = [":bundle"],
+    data = glob([
+        "*_golden.js_",
+    ]) + [
+        ":bundle",
+        ":bundle.min.js",
+    ],
     node_modules = "//internal/test:node_modules",
+    deps = [
+        "//internal/e2e:check_lib",
+    ],
 )
