+++ internal/e2e/rollup_code_splitting/bundle-min_golden.js_
@@ -1,0 +1,12 @@
+// SystemJS boilerplate/entry point TEMPLATE for code-split rollup_bundle.
+// GENERATED BY Bazel
+(function(global) {
+System.config({
+  packages: {
+    '': {map: {"./main1.js": "bundle.cs.min/main1.js", "./additional_entry.js": "bundle.cs.min/additional_entry.js"}, defaultExtension: 'js'},
+  }
+});
+System.import('main1.js').catch(function(err) {
+  console.error(err);
+});
+})(this);
